GameManager and EventHandler
Split the responsibilities of GameManager and EventHandler
-- Removed Game Status methods from EventHandler and moved to GameManger
-- EventHandler handles UI methods
-- Delete GameStateHandler
-- Returned shell code for enemies list
