+++ Balloonatic/Assets/Scripts/EventHandler.cs
@@ -19,19 +19,9 @@ public class EventHandler : MonoBehaviour
     private Camera cam;
     //[SerializeField] public GameObject timerObject;
 
-    [Header("Game Status")] 
-    public int health = 5;
-    public bool gameOver, gameActive, gamePaused;
-    private int levelScore, totalScore = 0;
-
-    //private GameManager gm = GameManager.Instance;
+    private GameManager gm = GameManager.Instance;
     //private Timer ts;
 
-
-    private void Start()
-    {
-        SetLevel();
-    }
 
     private void Update()
     {
@@ -41,21 +31,7 @@ public class EventHandler : MonoBehaviour
         }
     }
 
-    private void SetLevel()
-    {
-        Debug.Log("SETTING LVL");
-        cam = Camera.main;
-        EverythingFalse();
-        Time.timeScale = 1;
-        //timer = 0
-
-        UpdateScore(0);
-        UpdateHealth(0);
-
-        NewWave();
-    }
-
-    private void EverythingFalse()
+    public void UIFalse()
     {
         nxtCan.SetActive(false);
         ovrCan.SetActive(false);
@@ -64,51 +40,19 @@ public class EventHandler : MonoBehaviour
         loseCan.SetActive(false);
         pauseMen.SetActive(false);
         //ts.StartTime();
-
-        gameOver = false;
-        gameActive = true;
-        gamePaused = false;
     }
 
-    private void NewWave()
+    public void DisplayHealth(int hp)
     {
-        Debug.Log("New wave");
-        // incriment level count
-        // change walls or something
-        // ? some effect for enemies
+        healthText.text = "Lives: " + hp.ToString();
     }
 
-    private int CalcLevelScore(int score)
+    public void DisplayScore(int lvlScore)
     {
-        Debug.Log("Score Calc");
-        int newScore = score;
-        //float timeElapsed = ts.GetTime();
-        //float timeMult = (timeElapsed / 60 < 4) ? (1 - timeElapsed / 300) : 0.1f;
-
-        //newScore = Mathf.RoundToInt(score * ((1 + (float)health / fullHealth) * timeMult) - (enemyFactor * enemies.Count) * makeScoreBigger);
-        //  or multiply by (1 - enemies.Count / enemies2Spawn);
-        return Mathf.Max(0, newScore);
+        goalText.text = "Score: " + lvlScore.ToString();
     }
 
-    public void EndLevel(bool alive, int score)
-    {
-        Time.timeScale = 0;
-        gameActive = false;
-
-        totalScore += CalcLevelScore(score);
-
-        if (alive)
-        {
-            // wave intermission
-            NewWave();
-        }
-        else
-        {
-            LoseGame(totalScore);
-        }
-    }
-
-    private void CheckHS()
+    public void CheckHS(int totalScore)
     {
         cam.GetComponent<AudioSource>().Stop();
         ovrCan.SetActive(true);
@@ -133,13 +77,13 @@ public class EventHandler : MonoBehaviour
         hsInput.gameObject.SetActive(false);
     }
 
-    private void LoseGame(int finalScore)
+    public void LoseGame(int finalScore)
     {
-        gameOver = true;
+        gm.gameOver = true;
         loseCan.SetActive(true);
         loseScore.text = finalScore.ToString();
 
-        CheckHS();
+        CheckHS(finalScore);
         Cursor.lockState = CursorLockMode.None;
         Cursor.visible = true;
     }
@@ -157,12 +101,12 @@ public class EventHandler : MonoBehaviour
 
     public void PauseGame()
     {
-        if (gameOver) return;
-        if (gamePaused)
+        if (gm.gameOver) return;
+        if (gm.gamePaused)
         {
             pauseMen.SetActive(false);
-            gamePaused = false;
-            gameActive = true;
+            gm.gamePaused = false;
+            gm.gameActive = true;
             Time.timeScale = 1;
 
             //Cursor.lockState = CursorLockMode.Locked;
@@ -171,8 +115,8 @@ public class EventHandler : MonoBehaviour
         else
         {
             Time.timeScale = 0;
-            gameActive = false;
-            gamePaused = true;
+            gm.gameActive = false;
+            gm.gamePaused = true;
             pauseMen.SetActive(true);
 
             Cursor.lockState = CursorLockMode.None;
@@ -182,7 +126,7 @@ public class EventHandler : MonoBehaviour
 
     public void Restart()
     {
-        if (gamePaused)
+        if (gm.gamePaused)
         {
             LoadScene(SceneManager.GetActiveScene().buildIndex);
         }
@@ -196,28 +140,5 @@ public class EventHandler : MonoBehaviour
             Application.Quit();
         }
         else LoadScene(0);
-    }
-
-    public bool getGameActive()
-    {
-        return (gameActive && !gamePaused);
-    }
-
-    public void UpdateHealth(int change = -1)
-    {
-        health += change;
-        healthText.text = "Lives: " + health.ToString();
-
-        if (health <= 0)
-        {
-            health = 0;
-            EndLevel(false, levelScore);
-        }
-    }
-
-    public void UpdateScore(int change = 1)
-    {
-        levelScore += change;
-        goalText.text = "Score: " + levelScore.ToString();
     }
 }
