--- Balloonatic/Assets/Scripts/GameStateHandler.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1a569e3188dcb4441816ad70b989be8f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
